cuda-oxide-codegen: rustc-independent dialect-mir to PTX backend
Extract the post-translation pipeline tail out of mir-importer into a new
rustc-free crate, cuda-oxide-codegen. Its public boundary is dialect-mir: a
frontend assembles a dialect-mir module and calls
compile_to_ptx(context, module, options) to get PTX with no rustc per kernel
and no nvrtc.

The crate owns module verification, lowering to the LLVM dialect, LLVM-IR
export, architecture feature detection and PTX target selection, and the
opt/llc invocation. Every CUDA_OXIDE_* environment read inside the backend is
replaced by an explicit BackendOptions; mir-importer builds one at its own
boundary and re-exports the codegen surface, so rustc-codegen-cuda and
cargo-oxide are unchanged.

The Rust MIR path (run_pipeline) keeps its rustc-specific passes (mem2reg,
loop unroll, device externs, libdevice and NVVM linking) and calls the same
backend primitives, so the lowering pipeline exists once with no duplicate
copy. The fma-contraction policy (--no-fmad) threads through both the IR-level
contract flag and the llc -fp-contract gate via BackendOptions.

Public surface: register_backend_dialects, compile_to_ptx, PtxConfig,
PtxError, PipelineError, DeviceExternDecl, DeviceExternAttrs, DeviceExternType.
PtxConfig and PtxError are non_exhaustive; the run_pipeline plumbing items are
doc(hidden). The backend has no rustc linkage, so a consumer needs neither the
rustc_private feature nor a nightly matched to rustc_driver.
